service/requirements.txt
opencv-python
python-dotenv
torch
transformers
ultralytics
Pillow
requests
psycopg2-binary
